+++ superwall/src/main/java/com/superwall/sdk/misc/ActivityProvider.kt
@@ -1,0 +1,11 @@
+package com.superwall.sdk.misc
+
+import android.app.Activity
+
+interface ActivityProvider {
+    /**
+     * Provide an activity for Superwall to use instead of passing it in
+     * on configure.
+     */
+    fun getCurrentActivity(): Activity?
+}